Fish.js was deleted, as it was redundant to app.js. I also relocated FishSearch.js to components/search to make it more organized.
