--- fish-finder/frontend/src/Fish.js
@@ -1,12 +1,0 @@
-import React from 'react';
-import FishSearch from './FishSearch';
-
-function App() {
-    return (
-        <div className="App">
-            <FishSearch />
-        </div>
-    );
-}
-
-export default App;
